Azure Text Analytics/requirements.txt
azure.ai.textanalytics>=5.2.0
azure.identity>=1.5.0